feat: Implement ContributorCard component and update ContributorsDetails
Create a reusable ContributorCard component for displaying contributor information.
Refactor ContributorsDetails to use InfiniteMovingCards for horizontal scrolling of contributor cards.
